Resolve 64bit crashes

+++ src/usb-mic/audiodev-wasapi.cpp
@@ -898,12 +898,12 @@ static BOOL CALLBACK WASAPIDlgProc(HWND hW, UINT uMsg, WPARAM wParam, LPARAM lPa
 
 	switch (uMsg) {
 	case WM_CREATE:
-		SetWindowLongPtr(hW, GWLP_USERDATA, (LONG)lParam);
+		SetWindowLongPtr(hW, GWLP_USERDATA, lParam);
 		break;
 	case WM_INITDIALOG:
 	{
 		s = (WASAPISettings *)lParam;
-		SetWindowLongPtr(hW, GWLP_USERDATA, (LONG)lParam);
+		SetWindowLongPtr(hW, GWLP_USERDATA, lParam);
 		int buffering = 50;
 		LoadSetting(s->dev_type, s->port, APINAME, N_BUFFER_LEN_SRC, buffering);
